aggiunto alert su Sei sicuro di eliminare tutti i salvataggi all e singolo

+++ FrontEnd/WinningStrategy/src/app/components/home/salvati/salvati.component.ts
@@ -27,24 +27,26 @@ export class SalvatiComponent implements OnInit {
     }
 
     deleteSalvataggio(id: number) {
+      if (window.confirm('Sei sicuro di cancellare il salvataggio?')) {
         this.salvatiSrv.deleteSalvataggio(id).subscribe( () => {
           this.salvatiSrv.getSavedData().subscribe((data) => {
             console.log(data);
             this.savedDataList = data;
         });
         });
+      }
     }
 
     deleteSalvataggi() {
-        this.salvatiSrv.deleteSalvataggi(this.user?.idUtente).subscribe( () => {
-          this.salvatiSrv.getSavedData().subscribe((data) => {
-            console.log(data);
-
-            this.savedDataList = data;
-        });
-        });
-    }
-
+      if (window.confirm('Sei sicuro di cancellare tutti i salvataggi?')) {
+          this.salvatiSrv.deleteSalvataggi(this.user?.idUtente).subscribe(() => {
+            this.salvatiSrv.getSavedData().subscribe((data) => {
+              console.log(data);
+              this.savedDataList = data;
+            });
+          });
+      }
+  }
     goBack(): void {
       this.router.navigate(['/pitch']); 
     }
